+++ timelock-server/src/integTest/java/com/palantir/atlasdb/timelock/TestableTimelockCluster.java
@@ -23,6 +23,7 @@ import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import org.awaitility.Awaitility;
@@ -36,7 +37,7 @@ import com.palantir.atlasdb.timelock.util.TestProxies;
 import com.palantir.common.concurrent.PTExecutors;
 import com.palantir.lock.LockRefreshToken;
 import com.palantir.lock.LockService;
-import com.palantir.lock.v2.BridgingTimelockService;
+import com.palantir.lock.client.LeasingTimelockClient;
 import com.palantir.lock.v2.LockRequest;
 import com.palantir.lock.v2.LockResponse;
 import com.palantir.lock.v2.LockToken;
@@ -199,7 +200,7 @@ public class TestableTimelockCluster {
 
     public StartIdentifiedAtlasDbTransactionResponse startIdentifiedAtlasDbTransaction(
             StartIdentifiedAtlasDbTransactionRequest request) {
-        return rpcClient(defaultClient).startAtlasDbTransaction(request).toStartTransactionResponse();
+        return timelockRpcClient(defaultClient).startAtlasDbTransaction(request).toStartTransactionResponse();
     }
 
     public TimestampService timestampService() {
@@ -215,18 +216,18 @@ public class TestableTimelockCluster {
     }
 
     public TimelockService timelockServiceForClient(String client) {
-        return BridgingTimelockService.create(timelockRpcClient(client));
+        return LeasingTimelockClient.create(proxies.failoverForClient(client, TimelockRpcClient.class));
+    }
+
+    public <T> CompletableFuture<T> runWithRpcClientAsync(Function<TimelockRpcClient, T> function) {
+        return CompletableFuture.supplyAsync(() -> function.apply(timelockRpcClient()));
     }
 
     public TimelockRpcClient timelockRpcClient() {
         return timelockRpcClient(defaultClient);
     }
 
-    public TimelockRpcClient timelockRpcClient(String client) {
-        return proxies.failoverForClient(client, TimelockRpcClient.class);
-    }
-
-    private TimelockRpcClient rpcClient(String client) {
+    private TimelockRpcClient timelockRpcClient(String client) {
         return proxies.failoverForClient(client, TimelockRpcClient.class);
     }